Initial implementation of a geospatial agent
Initial implementaiton of a geospatial agent based on Amazon Bedrock provided
foundational model Claude 2.

In this version, we have  CLI that invokes the agent inside and outside a docker
container. It is a plan-and-solve agent that plans for a high level customer provided
task and writes code according to the plan.