[Bugfix] Returning generated code file name and vis file name through agent
Returning the generated code file and visualization file, may that be an html
or png was missed in the first release. I also simplified the `exec` usage
inside a single utility function.

As a bonus, I also noted that since this project uses Gen AI, it might require
one or more iterations to be successful.

test: Running the agent.